refactor(comments): extract ICommentsContext + split MasterModel comments into partial
Segregates the comment-provider concern (Disqus + Giscus config, provider
resolution, enablement predicates) out of IMasterModel/MasterModel:

- New ICommentsContext interface carries the 20 comment members.
- IMasterModel : IPublishedContent, ICommentsContext (composes it; all existing
  members still resolve via inheritance — purely additive, no break to v5
  consumers or external IMasterModel implementers).
- New MasterModel.Comments.cs partial implements the concern; MasterModel is
  now partial and stripped of the comment code (~145 lines moved). Ctor wiring
  unchanged (CommentsOptions still assigned in MasterModel ctors; partial
  classes share members).
- Caches IsDisqusEnabled and CommentsProvider behind nullable backing fields
  (were expression-bodied, recomputed each access) to avoid re-resolving
  Fallback.ToAncestors + the all-or-nothing override on repeat access. Static
  ResolveProvider/ResolveGiscusRequired helpers move to the partial, remain
  internal static (tests unaffected).

Shipped comment partials are intentionally NOT narrowed — they keep
@model PostModel for backwards compatibility with v5/custom-theme overrides.
ICommentsContext is organizational (ISP/future-use); IMasterModel still
exposes every comment member transitively.

Builds clean both lanes; 209/209 tests pass.
