+++ src/Articulate/Models/IMasterModel.cs
@@ -1,5 +1,4 @@
 #nullable enable
-using Articulate.Options;
 using Umbraco.Cms.Core.Models.PublishedContent;
 
 namespace Articulate.Models
@@ -7,7 +6,7 @@ namespace Articulate.Models
     /// <summary>
     /// Represents the master model for Articulate pages.
     /// </summary>
-    public interface IMasterModel : IPublishedContent
+    public interface IMasterModel : IPublishedContent, ICommentsContext
     {
         /// <summary>
         /// Gets the current theme.
@@ -60,62 +59,6 @@ namespace Articulate.Models
         /// Gets the number of items per page.
         /// </summary>
         public int PageSize { get; }
-
-        /// <summary>
-        /// Gets the Disqus short name.
-        /// </summary>
-        public string DisqusShortName { get; }
-
-        /// <summary>
-        /// Gets the comment options (appsettings defaults) resolved for this model.
-        /// Per-blog overrides still come from the document via <c>Fallback.ToAncestors</c>;
-        /// this carries only the global appsettings fallback values.
-        /// </summary>
-        public ArticulateCommentsOptions CommentsOptions { get; }
-
-        /// <summary>
-        /// Gets whether Disqus comments are enabled and configured with a valid shortname.
-        /// </summary>
-        public bool IsDisqusEnabled { get; }
-
-        /// <summary>
-        /// Gets the resolved comment provider.
-        /// </summary>
-        public ArticulateConstants.Comments.Provider CommentsProvider { get; }
-
-        /// <summary>
-        /// Gets whether any comment provider is configured.
-        /// </summary>
-        public bool IsCommentsEnabled { get; }
-
-        public string GiscusScriptSrc { get; }
-
-        public string GiscusRepo { get; }
-
-        public string GiscusRepoId { get; }
-
-        public string GiscusCategory { get; }
-
-        public string GiscusCategoryId { get; }
-
-        public string GiscusMapping { get; }
-
-        public string GiscusStrict { get; }
-
-        public string GiscusReactionsEnabled { get; }
-
-        public string GiscusEmitMetadata { get; }
-
-        public string GiscusInputPosition { get; }
-
-        public string GiscusTheme { get; }
-
-        public string GiscusLang { get; }
-
-        public string GiscusLoading { get; }
-
-        public bool IsGiscusEnabled { get; }
-
         /// <summary>
         /// Gets the custom RSS feed URL.
         /// </summary>
